requirements.txt
# Core ML and Data Processing
transformers>=4.30.0
datasets>=2.10.0
torch>=2.0.0
accelerate>=0.26.0
bitsandbytes>=0.42.0
peft>=0.10.0
numpy>=1.24.0
sentence-transformers>=2.2.2

# Data sources
edgartools>=0.1.0

# llm providers
openai>=1.0.0
anthropic>=0.39.0

# Data collection
crawl4ai>=0.7.0

# Text Processing and Utilities
unidecode>=1.3.0
tqdm>=4.65.0
bs4>=0.0.1
pandas>=2.0.0
pdfminer.six>=20221105

# Environment and Configuration
python-dotenv>=1.0.0

# Other installs
# playwright install
